requirements.txt
asgiref
attrs  
backoff
blinker
branca
Brotli
certifi
cffi
charset-normalizer
crispy-bootstrap5
cryptography
dj-database-url
Django
django-crispy-forms
django-environ
django-filter
djangorestframework
folium
h11
h2
hpack
hyperframe
idna
Jinja2
kaitaistruct
Markdown
MarkupSafe
MouseInfo
mysql-connector
mysql-connector-python
numpy
opencage
outcome
packaging
phonenumbers
pillow
psycopg2
psycopg2-binary
pyasn1
PyAutoGUI
pycparser
pydivert
PyGetWindow
PyMsgBox
pyOpenSSL
pyparsing
pyperclip
PyRect
PyScreeze
PySocks
python-dotenv
pytweening
requests
selenium-wire
sniffio
sortedcontainers
sqlparse
trio
trio-websocket
typing_extensions
tzdata
urllib3
urllib3
webdriver-manager
wsproto
xyzservices
zstandard